fixed dtypes to work in numba and interpreted python

+++ pyflwdir/core_d8.py
@@ -69,7 +69,7 @@ def from_array(flwdir):
     idxs_ds = np.full(n, core._mv, dtype=np.uint32)
     # loop over sparse indices
     for idx0 in idxs_dense:
-        i = idxs_sparse[idx0]
+        i = np.uint32(idxs_sparse[idx0])
         dr, dc = drdc(flwdir_flat[idx0])
         r = int(idx0 // ncol + dr) 
         c = int(idx0  % ncol + dc)
@@ -80,8 +80,8 @@ def from_array(flwdir):
             idxs_ds[i] = i
             pits_lst.append(np.uint32(i))
         else:        
-            idx_ds = idx0 + dc + dr * ncol
-            i_ds = idxs_sparse[idx_ds]
+            idx_ds = np.intp(idx0 + dc + dr * ncol)
+            i_ds = np.uint32(idxs_sparse[idx_ds])
             if i_ds == core._mv:
                 raise ValueError('invalid D8 data')
             idxs_ds[i] = i_ds
